update my fork

+++ package/lean/luci-app-unblockmusic/root/usr/share/UnblockNeteaseMusic/update_core.sh
@@ -1,4 +1,4 @@
-
+#!/bin/sh
 
 function check_if_already_running(){
 	running_tasks="$(ps |grep "unblockneteasemusic" |grep "update_core" |grep -v "grep" |awk '{print $1}' |wc -l)"
@@ -24,6 +24,7 @@ function check_latest_version(){
 		else
 			echo -e "\nLocal version: $(cat /usr/share/UnblockNeteaseMusic/local_ver 2>/dev/null), cloud version: ${latest_ver}." >>/tmp/unblockmusic_update.log
 			echo -e "You're already using the latest version." >>/tmp/unblockmusic_update.log
+			[ "${luci_update}" == "n" ] && /etc/init.d/unblockmusic restart
 			exit 3
 		fi
 	fi
@@ -37,7 +38,9 @@ function update_core(){
 
 	wget-ssl --no-check-certificate -t 1 -T 10 -O  /tmp/unblockneteasemusic/core/core.tar.gz "https://github.com/nondanee/UnblockNeteaseMusic/archive/master.tar.gz"  >/dev/null 2>&1
 	tar -zxf "/tmp/unblockneteasemusic/core/core.tar.gz" -C "/tmp/unblockneteasemusic/core/" >/dev/null 2>&1
-	rm -f /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/ca.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.key
+	if [ -e "/usr/share/UnblockNeteaseMusic/ca.crt" ] && [ -e "/usr/share/UnblockNeteaseMusic/server.crt" ] && [ -e "/usr/share/UnblockNeteaseMusic/server.key" ] ; then
+		rm -f /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/ca.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.key
+	fi
 	cp -a /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/* "/usr/share/UnblockNeteaseMusic/"
 	rm -rf "/tmp/unblockneteasemusic" >/dev/null 2>&1
 
@@ -45,14 +48,14 @@ function update_core(){
 		echo -e "Failed to download core." >>/tmp/unblockmusic_update.log
 		exit 1
 	else
-		[ "${luci_update}" == "y" ] && touch "/usr/share/unblockneteasemusic/update_successfully"
 		echo -e "${latest_ver}" > /usr/share/UnblockNeteaseMusic/local_ver
-		/etc/init.d/unblockmusic restart
+		cat /usr/share/UnblockNeteaseMusic/package-lock.json | grep version |awk -F ':' '{print $2}' | cut -c3-8 > /usr/share/UnblockNeteaseMusic/core_ver
 	fi
 
 	echo -e "Succeeded in updating core." >/tmp/unblockmusic_update.log
 	echo -e "Local version: $(cat /usr/share/UnblockNeteaseMusic/local_ver 2>/dev/null), cloud version: ${latest_ver}.\n" >>/tmp/unblockmusic_update.log
-	node /usr/share/UnblockNeteaseMusic/app.js -v > /usr/share/UnblockNeteaseMusic/core_ver
+	
+	/etc/init.d/unblockmusic restart
 }
 
 function main(){
@@ -60,5 +63,7 @@ function main(){
 	check_latest_version
 }
 
+  luci_update="n"
 	[ "$1" == "luci_update" ] && luci_update="y"
 	main
+
